refactor: 채팅방 리스트/생성 컴포넌트를 Container/Presentational 패턴으로 분리
