+++ src/components/chat/ChatRoomForm.jsx
@@ -1,72 +1,18 @@
-import { useEffect, useState } from "react";
-import { useNavigate } from "react-router-dom";
-import { useSelector } from "react-redux";
-import { useCreateChatRoom } from "hooks/chat/useCreateChatRoom";
-import LoadingSpinner from "components/common/LoadingSpinner.jsx";
 import ErrorMessage from "components/common/ErrorMessage";
 import FormSubmitButton from "components/common/FormSubmitButton";
 import { Lightbulb } from "lucide-react";
 
-export default function ChatRoomForm() {
-  const [title, setTitle] = useState("");
-  const [description, setDescription] = useState("");
-  const [category, setCategory] = useState("");
-
-  const user = useSelector((state) => state.user.user);
-  const isLoading = useSelector((state) => state.user.isLoading);
-
-  const navigate = useNavigate();
-  const { mutateAsync, isPending } = useCreateChatRoom(navigate);
-
-  const [errors, setErrors] = useState({});
-  const [didAlert, setDidAlert] = useState(false);
-
-  useEffect(() => {
-    // 로딩이 끝났고, 로그인하지 않은 경우에만 실행
-    if (!isLoading && !user && !didAlert) {
-      alert("로그인이 필요합니다.");
-      setDidAlert(true);
-      navigate("/login", { state: { from: "/chat/create" } });
-    }
-  }, [user, isLoading, navigate, didAlert]);
-
-  const validateForm = () => {
-    const newErrors = {};
-    if (!title.trim()) newErrors.title = "제목을 입력해주세요.";
-    if (!description.trim()) newErrors.description = "설명을 입력해주세요.";
-    if (!category) newErrors.category = "카테고리를 선택해주세요.";
-    if (!user) newErrors.user = "로그인이 필요합니다.";
-
-    setErrors(newErrors);
-    return Object.keys(newErrors).length === 0;
-  };
-
-  const handleChatRoomFormSubmit = async (e) => {
-    e.preventDefault();
-
-    if (!validateForm()) return;
-
-    try {
-      await mutateAsync({
-        title,
-        description,
-        category,
-        user,
-      });
-    } catch (err) {
-      console.error("채팅방 생성 오류:", err);
-      if (err.message.includes("제목과 사용자 ID는 필수입니다")) {
-        alert("필수 정보가 누락되었습니다. 제목과 로그인 정보를 확인해주세요.");
-      } else if (err.code === "permission-denied") {
-        alert("채팅방 생성 권한이 없습니다. 로그인 상태를 확인해주세요.");
-      } else {
-        alert("채팅방 생성 중 오류가 발생했어요. 잠시 후 다시 시도해주세요.");
-      }
-    }
-  };
-
-  if (isLoading || !user) return <LoadingSpinner />;
-
+export default function ChatRoomForm({
+  title,
+  setTitle,
+  description,
+  setDescription,
+  category,
+  setCategory,
+  errors,
+  handleSubmit,
+  isSubmitting,
+}) {
   return (
     <article
       className="relative min-h-screen bg-cover bg-center bg-no-repeat"
@@ -88,7 +34,7 @@ export default function ChatRoomForm() {
 
         {/* 채팅방 생성 폼 */}
         <form
-          onSubmit={handleChatRoomFormSubmit}
+          onSubmit={handleSubmit}
           className="mt-12 space-y-6 bg-white/60 backdrop-blur-lg p-10 rounded-3xl shadow-md border border-white/30 text-gray-800"
         >
           {/* 제목 */}
@@ -146,7 +92,7 @@ export default function ChatRoomForm() {
 
           {/* 전송 버튼 */}
           <div className="flex justify-end">
-            <FormSubmitButton isLoading={isPending} label="채팅방 만들기" />
+            <FormSubmitButton isLoading={isSubmitting} label="채팅방 만들기" />
           </div>
         </form>
