+++ src/pages/chat/Create.jsx
@@ -1,9 +1,9 @@
-import ChatRoomForm from "components/chat/ChatRoomForm";
+import ChatRoomFormContainer from "containers/chat/ChatRoomFormContainer";
 
 export default function CreateChatRoomPage() {
   return (
     <div className="pt-16">
-      <ChatRoomForm />
+      <ChatRoomFormContainer />
     </div>
   );
 }
